requirements.txt
anthropic>=0.40.0
ddgs>=0.1.0
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiofiles>=23.0.0
pydantic>=2.0.0
